requirements.txt
click==3.3
gevent==1.0.1
pyzmq==14.4.1